+++ client/src/components/supplier/InactiveSupplierTable.tsx
@@ -1,24 +1,20 @@
-import { Input, Table, Button, Space, Popconfirm, Modal, Tag } from "antd";
+import { Input, Table, Button, Space } from "antd";
 import React, { useState } from "react";
 import { useAllSuppliers} from "../../services/queries/supplierQueries";
 import { SupplierInfo } from "../../../../shared/types/Supplier";
-import { useDeleteSupplier } from "../../services/mutations/supplierMutation";
-import SupplierForm from "./SupplierForm";
-
+import { ColumnsType } from "antd/es/table";
+import {useAuth} from "../../context/AuthContext"
 
 const ListTable = () => {
   
   const [searchValue, setSearchValue] = useState("");
-  const [editRecord, setEditRecord] = useState<any>(null);
-  const [editModalVisible, setEditModalVisible] = useState(false);
-
+  const {user} = useAuth()
   const allSuppliersQuery = useAllSuppliers();
-  const deleteSupplierMutation = useDeleteSupplier();
   console.log("All Suppliers Query:", allSuppliersQuery.data);
 
   const Source = allSuppliersQuery.data
   ? allSuppliersQuery.data
-      .filter((queryResult: SupplierInfo) => queryResult.status === 'inactive') // Only include suppliers with status === 'active'
+      .filter((queryResult: SupplierInfo) => queryResult.status === 'inactive') 
       .map((queryResult: SupplierInfo) => {
         return {
           key: queryResult._id,
@@ -27,7 +23,7 @@ const ListTable = () => {
           mobileNumber: queryResult.mobileNumber,
           email: queryResult.email,
           address: queryResult.address,
-          status: queryResult.status, // Add status field to the source
+          status: queryResult.status, 
         };
       })
   : [];
@@ -38,19 +34,13 @@ const ListTable = () => {
     setSearchValue(value);
   };
 
-  const handleDelete = (key: string) => {
-    console.log("Delete Key:", key);
-    deleteSupplierMutation.mutate(key);
-    // const filteredDataSource = dataSource.filter(item => item.key !== key);
-    // setDataSource(filteredDataSource);
-  };
   const handleActivate = (key: string) => {
     console.log("Activate Key:");
    
   }
 
 
-  const columns = [
+  const columns: ColumnsType = [
     {
       title: "SID",
       dataIndex: "sid",
@@ -76,12 +66,17 @@ const ListTable = () => {
       dataIndex: "address",
       key: "address",
     },
-    {
+  ];
+  if (user?.role === "invmanager") {
+    columns.push({
       title: "Action",
       key: "action",
       render: (text: string, record: SupplierInfo) => (
         <Space size="middle">
-          <Button type="primary" onClick={() => handleActivate(record.sid || "")}>
+          <Button
+            type="primary"
+            onClick={() => handleActivate(record.sid || "")}
+          >
             Activate
           </Button>
           {/* <Button type="dashed" onClick={() => handleDelete(record.sid || "")}>Delete</Button> */}
@@ -95,8 +90,8 @@ const ListTable = () => {
           </Popconfirm> */}
         </Space>
       ),
-    },
-  ];
+    });
+  }
 
 
   return (
